pyproject.toml
[tool.poetry]
name = "greenbox"
version = "0.1.0"
description = "Raspberry Pi IoT device for plant growing environment monitoring and control"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Home Automation",
]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9"
adafruit-circuitpython-dht = ">=3.7.0"
schedule = ">=1.1.0"
PyYAML = ">=6.0"

[tool.poetry.group.rpi.dependencies]
RPi-GPIO = {version = ">=0.7.0", markers = "platform_machine == 'armv7l' or platform_machine == 'aarch64'"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"
black = ">=23.0.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"
poethepoet = "^0.33.1"

[tool.poetry.scripts]
greenbox = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 79
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79

[tool.poe.tasks]
lint = [
    { cmd = "black src tests" },
    { cmd = "isort src tests" },
    { cmd = "flake8 src tests" }
]
test = "pytest tests"
coverage = "pytest --cov=src tests/"
format = [
    { cmd = "black src tests" },
    { cmd = "isort src tests" }
]